added last test and refactoring and added allure-report
