+++ diplom2-2/allure-results/0b4c8439-9446-4754-839d-ea950f800ab7-result.json
@@ -1,0 +1,1 @@
+{"uuid":"0b4c8439-9446-4754-839d-ea950f800ab7","historyId":"c962ea1312bb3c6da18adc2e679fa13e","fullName":"CreateOrderTest.createOrderWithoutAuthTest","labels":[{"name":"package","value":""},{"name":"testClass","value":"CreateOrderTest"},{"name":"testMethod","value":"createOrderWithoutAuthTest"},{"name":"suite","value":"CreateOrderTest"},{"name":"host","value":"kryabova-home"},{"name":"thread","value":"271252@kryabova-home.main(1)"},{"name":"framework","value":"junit4"},{"name":"language","value":"java"}],"links":[],"name":"createOrderWithoutAuthTest","status":"passed","stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1663538488207,"stop":1663538488302}
