+++ diplom2-2/allure-results/0ea791ec-1c80-4435-ba37-10acac6077dd-result.json
@@ -1,0 +1,1 @@
+{"uuid":"0ea791ec-1c80-4435-ba37-10acac6077dd","historyId":"71b694d34b957a69d7568a7136511f52","fullName":"CreateOrderTest.createOrderWithoutIngredientsTest","labels":[{"name":"package","value":""},{"name":"testClass","value":"CreateOrderTest"},{"name":"testMethod","value":"createOrderWithoutIngredientsTest"},{"name":"suite","value":"CreateOrderTest"},{"name":"host","value":"kryabova-home"},{"name":"thread","value":"271252@kryabova-home.main(1)"},{"name":"framework","value":"junit4"},{"name":"language","value":"java"}],"links":[],"name":"createOrderWithoutIngredientsTest","status":"passed","stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1663538489185,"stop":1663538489289}
