+++ diplom2-2/allure-results/344de466-2aac-43ee-a2f4-7c5915fec2bb-result.json
@@ -1,0 +1,1 @@
+{"uuid":"344de466-2aac-43ee-a2f4-7c5915fec2bb","historyId":"cb44f6a8e555b93bf3e82160a244026b","fullName":"CreateOrderTest.createOrderWithNotValidIngredientHashTest","labels":[{"name":"package","value":""},{"name":"testClass","value":"CreateOrderTest"},{"name":"testMethod","value":"createOrderWithNotValidIngredientHashTest"},{"name":"suite","value":"CreateOrderTest"},{"name":"host","value":"kryabova-home"},{"name":"thread","value":"271252@kryabova-home.main(1)"},{"name":"framework","value":"junit4"},{"name":"language","value":"java"}],"links":[],"name":"createOrderWithNotValidIngredientHashTest","status":"passed","stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1663538488304,"stop":1663538489183}
