+++ diplom2-2/allure-results/4ea93557-e099-4fd6-8df7-137718e2884a-result.json
@@ -1,0 +1,1 @@
+{"uuid":"4ea93557-e099-4fd6-8df7-137718e2884a","historyId":"7317014d0ae6ac5693ec8e50c963d87e","fullName":"ChangeUserDataTest.changeAuthUserExistEmailTest","labels":[{"name":"package","value":""},{"name":"testClass","value":"ChangeUserDataTest"},{"name":"testMethod","value":"changeAuthUserExistEmailTest"},{"name":"suite","value":"ChangeUserDataTest"},{"name":"host","value":"kryabova-home"},{"name":"thread","value":"271252@kryabova-home.main(1)"},{"name":"framework","value":"junit4"},{"name":"language","value":"java"}],"links":[],"name":"changeAuthUserExistEmailTest","status":"passed","stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1663538480200,"stop":1663538482442}
