+++ diplom2-2/allure-results/5b2fc1c9-41a7-430a-9fef-efc42fd0ff7c-result.json
@@ -1,0 +1,1 @@
+{"uuid":"5b2fc1c9-41a7-430a-9fef-efc42fd0ff7c","historyId":"8be95c303cac7f183f11b33eb92703ec","fullName":"LoginUserTest.singInWithValidData","labels":[{"name":"package","value":""},{"name":"testClass","value":"LoginUserTest"},{"name":"testMethod","value":"singInWithValidData"},{"name":"suite","value":"LoginUserTest"},{"name":"host","value":"kryabova-home"},{"name":"thread","value":"271252@kryabova-home.main(1)"},{"name":"framework","value":"junit4"},{"name":"language","value":"java"}],"links":[],"name":"singInWithValidData","status":"passed","stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1663538497976,"stop":1663538498952}
