+++ diplom2-2/allure-results/6724f1b6-e6e7-4c35-bc85-c5335a051a42-result.json
@@ -1,0 +1,1 @@
+{"uuid":"6724f1b6-e6e7-4c35-bc85-c5335a051a42","historyId":"717bd037955150a1941dd2312718c921","fullName":"CreateOrderTest.createOrderWithAuthTest","labels":[{"name":"package","value":""},{"name":"testClass","value":"CreateOrderTest"},{"name":"testMethod","value":"createOrderWithAuthTest"},{"name":"suite","value":"CreateOrderTest"},{"name":"host","value":"kryabova-home"},{"name":"thread","value":"271252@kryabova-home.main(1)"},{"name":"framework","value":"junit4"},{"name":"language","value":"java"}],"links":[],"name":"createOrderWithAuthTest","status":"passed","stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1663538487144,"stop":1663538488206}
