+++ diplom2-2/allure-results/725b8eb1-ce73-48e1-b71e-ad0eb582023f-result.json
@@ -1,0 +1,1 @@
+{"uuid":"725b8eb1-ce73-48e1-b71e-ad0eb582023f","historyId":"3bb35d729266b10d8966e72772e537ce","fullName":"ChangeUserDataTest.changeNotAuthUserTest","labels":[{"name":"package","value":""},{"name":"testClass","value":"ChangeUserDataTest"},{"name":"testMethod","value":"changeNotAuthUserTest"},{"name":"suite","value":"ChangeUserDataTest"},{"name":"host","value":"kryabova-home"},{"name":"thread","value":"271252@kryabova-home.main(1)"},{"name":"framework","value":"junit4"},{"name":"language","value":"java"}],"links":[],"name":"changeNotAuthUserTest","status":"passed","stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1663538483635,"stop":1663538484233}
