+++ diplom2-2/allure-results/8fbb99c8-b0ae-433b-a917-75cc73c111eb-result.json
@@ -1,0 +1,1 @@
+{"uuid":"8fbb99c8-b0ae-433b-a917-75cc73c111eb","historyId":"adf33e6b566fdc84ff749d42178b3e2d","fullName":"CreateUserTest.createUserWithoutOneAttribute","labels":[{"name":"package","value":""},{"name":"testClass","value":"CreateUserTest"},{"name":"testMethod","value":"createUserWithoutOneAttribute"},{"name":"suite","value":"CreateUserTest"},{"name":"host","value":"kryabova-home"},{"name":"thread","value":"271252@kryabova-home.main(1)"},{"name":"framework","value":"junit4"},{"name":"language","value":"java"}],"links":[],"name":"createUserWithoutOneAttribute","status":"passed","stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1663538493442,"stop":1663538494341}
