+++ diplom2-2/allure-results/ad2727d1-59a4-4d88-b211-ac5da15657ca-result.json
@@ -1,0 +1,1 @@
+{"uuid":"ad2727d1-59a4-4d88-b211-ac5da15657ca","historyId":"fd435489588bb4d1373cafc66d8e45a8","fullName":"LoginUserTest.singInWithNotValidData","labels":[{"name":"package","value":""},{"name":"testClass","value":"LoginUserTest"},{"name":"testMethod","value":"singInWithNotValidData"},{"name":"suite","value":"LoginUserTest"},{"name":"host","value":"kryabova-home"},{"name":"thread","value":"271252@kryabova-home.main(1)"},{"name":"framework","value":"junit4"},{"name":"language","value":"java"}],"links":[],"name":"singInWithNotValidData","status":"passed","stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1663538497849,"stop":1663538497973}
