+++ diplom2-2/allure-results/b3ab32f1-b431-4541-b681-15cb676b3b6d-result.json
@@ -1,0 +1,1 @@
+{"uuid":"b3ab32f1-b431-4541-b681-15cb676b3b6d","historyId":"71333139a7ac44abe9702c8d43c3e135","fullName":"GettingOrdersTest.getNotAuthUserOrders","labels":[{"name":"package","value":""},{"name":"testClass","value":"GettingOrdersTest"},{"name":"testMethod","value":"getNotAuthUserOrders"},{"name":"suite","value":"GettingOrdersTest"},{"name":"host","value":"kryabova-home"},{"name":"thread","value":"271252@kryabova-home.main(1)"},{"name":"framework","value":"junit4"},{"name":"language","value":"java"}],"links":[],"name":"getNotAuthUserOrders","status":"passed","stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1663538495210,"stop":1663538495828}
