+++ diplom2-2/allure-results/bb918dfc-a171-46e2-a5c0-238d7c8c6446-result.json
@@ -1,0 +1,1 @@
+{"uuid":"bb918dfc-a171-46e2-a5c0-238d7c8c6446","historyId":"c5631b1035cc2cba714cb79086d55539","fullName":"CreateUserTest.createRepeatUser","labels":[{"name":"package","value":""},{"name":"testClass","value":"CreateUserTest"},{"name":"testMethod","value":"createRepeatUser"},{"name":"suite","value":"CreateUserTest"},{"name":"host","value":"kryabova-home"},{"name":"thread","value":"271252@kryabova-home.main(1)"},{"name":"framework","value":"junit4"},{"name":"language","value":"java"}],"links":[],"name":"createRepeatUser","status":"passed","stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1663538490939,"stop":1663538491810}
