+++ diplom2-2/allure-results/cad8c553-c82e-4653-ad22-2a6147ce0626-result.json
@@ -1,0 +1,1 @@
+{"uuid":"cad8c553-c82e-4653-ad22-2a6147ce0626","historyId":"a99caca21cf569bd825470e65065ebab","fullName":"CreateUserTest.createUserWithoutAttr","labels":[{"name":"package","value":""},{"name":"testClass","value":"CreateUserTest"},{"name":"testMethod","value":"createUserWithoutAttr"},{"name":"suite","value":"CreateUserTest"},{"name":"host","value":"kryabova-home"},{"name":"thread","value":"271252@kryabova-home.main(1)"},{"name":"framework","value":"junit4"},{"name":"language","value":"java"}],"links":[],"name":"createUserWithoutAttr","status":"passed","stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1663538491812,"stop":1663538493440}
