+++ diplom2-2/allure-results/cc3a5a9b-d08e-49b5-9187-cbab64ca2cd1-result.json
@@ -1,0 +1,1 @@
+{"uuid":"cc3a5a9b-d08e-49b5-9187-cbab64ca2cd1","historyId":"38553f7088fb9336f31b42696838fc5f","fullName":"ChangeUserDataTest.changeEmailAuthUserTest","labels":[{"name":"package","value":""},{"name":"testClass","value":"ChangeUserDataTest"},{"name":"testMethod","value":"changeEmailAuthUserTest"},{"name":"suite","value":"ChangeUserDataTest"},{"name":"host","value":"kryabova-home"},{"name":"thread","value":"271252@kryabova-home.main(1)"},{"name":"framework","value":"junit4"},{"name":"language","value":"java"}],"links":[],"name":"changeEmailAuthUserTest","status":"passed","stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1663538482622,"stop":1663538483633}
