+++ diplom2-2/allure-results/eeb1a90a-5be5-4d7c-bde2-65d7d04d86f6-result.json
@@ -1,0 +1,1 @@
+{"uuid":"eeb1a90a-5be5-4d7c-bde2-65d7d04d86f6","historyId":"c79a71e3e4f0282455801e15fca8abce","fullName":"GettingOrdersTest.getAuthUserOrders","labels":[{"name":"package","value":""},{"name":"testClass","value":"GettingOrdersTest"},{"name":"testMethod","value":"getAuthUserOrders"},{"name":"suite","value":"GettingOrdersTest"},{"name":"host","value":"kryabova-home"},{"name":"thread","value":"271252@kryabova-home.main(1)"},{"name":"framework","value":"junit4"},{"name":"language","value":"java"}],"links":[],"name":"getAuthUserOrders","status":"passed","stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1663538495829,"stop":1663538496508}
