+++ diplom2-2/allure-results/ffb7c1df-e520-40b1-b78d-f97df7b77276-result.json
@@ -1,0 +1,1 @@
+{"uuid":"ffb7c1df-e520-40b1-b78d-f97df7b77276","historyId":"3e9f0fb61d56fa02294412e7ecdfcec5","fullName":"ChangeUserDataTest.changeNameAuthUserTest","labels":[{"name":"package","value":""},{"name":"testClass","value":"ChangeUserDataTest"},{"name":"testMethod","value":"changeNameAuthUserTest"},{"name":"suite","value":"ChangeUserDataTest"},{"name":"host","value":"kryabova-home"},{"name":"thread","value":"271252@kryabova-home.main(1)"},{"name":"framework","value":"junit4"},{"name":"language","value":"java"}],"links":[],"name":"changeNameAuthUserTest","status":"passed","stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1663538484235,"stop":1663538484939}
